[babel-plugin] Move implementation detail tests
Moves files that test implementation details rather than the behavior of
the babel-plugin. These unit tests are now co-located with their code.
