+++ packages/babel-plugin/src/utils/__tests__/state-manager-test.js
