+++ packages/babel-plugin/src/utils/evaluation-errors.js
@@ -33,8 +33,8 @@ Could not determine the property being accessed.
 `;
 
 export const IMPORT_PATH_RESOLUTION_ERROR = `Could not resolve the path to the imported file.
-Please ensure that the theme file has a .stylex.js or .stylex.ts file extension and follows the
-rules for defining variariables: 
+Please ensure that the theme file has a .stylex.js or .stylex.ts extension and follows the
+rules for defining variables:
 
 https://stylexjs.com/docs/learn/theming/defining-variables/#rules-when-defining-variables
 `;
